project compilation and wasm execution
